pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[tool.pdm]
[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.3.1",
    "pytest-cov>=4.0.0",
    "coverage>=7.2.5",
]
qa = [
    "ruff>=0.0.263",
    "black>=23.3.0",
    "pre-commit>=3.2.2",
    "isort>=5.12.0",
    "pre-commit-hooks>=4.4.0",
    "pyright>=1.1.304",
    "nox>=2023.4.22",
]
docs = [
    "mkdocs-material>=9.1.8",
    "mkdocstrings-python>=0.9.0",
    "mkdocs-table-reader-plugin>=2.0",
]


[project]
name = "autoencoders"
version = "0.1.0"
description = "Experimenting with various autoencoder architectures."
readme = "README.md"
license = { text ="MIT License" }
requires-python = ">=3.10"

authors = [
    { name="Chris Santiago", email="cjsantiago@gatech.edu"}
]

dependencies = [
    "torch>=2.0.1",
    "hydra-core>=1.3.2",
    "pytorch-lightning>=2.0.7",
    "torchmetrics>=1.0.3",
    "torchvision>=0.15.2",
    "rich>=13.5.2",
    "wandb>=0.15.8",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.2",
    "hydra-joblib-launcher>=1.2.0",
]


[project.urls]
"Sourcecode" = "https://github.com/chris-santiago/autoencoders"
"Documentation" = "https://chris-santiago.github.io/autoencoders/"


[tool.setuptools]
include-package-data = true
zip-safe = false


[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests/*"]


[tool.check-manifest]
ignore = [
    "docs",
    "docs/**",
    "notebooks",
    "notebooks/**",
    ".pre-commit-config.yaml",
    ".gitignore",
    "noxfile.py",
    "mkdocs.yml",
]


[tool.isort]
profile = "black"


[tool.black]
line-length = 100


[tool.pytest.ini_options]
testpaths = ["tests/"]
addopts = "--cov=autoencoders --cov-report=html --verbose"


[tool.ruff]
line-length = 100
